configuration options for all features
Options can be toggled on/off globally, most of them can be restricted to a certain minimum MGMT power level. Powers above that number will have that feature enabled.

Some features do not have the MGMT power settings due to their nature, mostly because it would be more difficult and tricky to implement.